refactor - realoca funções de delene e edit das tooltips
